Ferme la recherche quand la page change
